+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- capability to fill gaps in consensus sequence with a designated character 
+  (e.g. 'N') instead of content from a reference sequence.
+- option `-r` in `medaka_consensus` to set the designated fill character.
+- option `--fill_char` in `medaka stitch` to set the designated fill character.
+
 ## [v1.6.1]
 ### Added
 - New models for R10.4.1 E8.2 400bps based sequencing chemistries.